move makeGaiaAssociationToken to index to try and fix circleci
